2008-08-01 09:21 UTC+0100 Miguel Angel Marchuet <miguelangel@marchuet.net>   * source/rdd/hbsix/sxcompr.c     * pacified warnings   * harbour/contrib/hbmzip/zip.c     ! set valid ZIP signature in header file (2.0)   * contrib/hbmzip/hbmzip.c   * contrib/hbmzip/readme.txt     * changed parameters of HB_UnzipFileInfo()       Now in 10-th parameter is @lCrypted and 11-th is @cComment     ! added "b" to fopen() flags to force binary mode in all systems       during CRC32 calculation   * harbour/contrib/hbmzip/tests/myunzip.prg     + show archive and file comment     + show encrypted files ('*' after file name)
    synced with harbour

+++ xharbour/source/rdd/hbsix/sxcompr.c
@@ -1,5 +1,5 @@
 /*
- * $Id: sxcompr.c,v 1.9 2007/07/05 11:09:09 marchuet Exp $
+ * $Id: sxcompr.c,v 1.10 2008/01/15 10:13:44 marchuet Exp $
  */
 
 /*
@@ -529,15 +529,15 @@ static ULONG hb_LZSSxEncode( PHB_LZSSX_COMPR pCompr )
          pCompr->ring_buffer[ s ] = ( BYTE ) c;
          if( s < MAXLENGTH - 1 )
             pCompr->ring_buffer[ s + RBUFLENGTH ] = ( BYTE ) c;
-         s = RBUFINDEX( s + 1 );
-         r = RBUFINDEX( r + 1 );
+         s = ( short int ) RBUFINDEX( s + 1 );
+         r = ( short int ) RBUFINDEX( r + 1 );
          hb_LZSSxNodeInsert( pCompr, r );
       }
       while( i++ < last_match_length )
       {
          hb_LZSSxNodeDelete( pCompr, s );
-         s = RBUFINDEX( s + 1 );
-         r = RBUFINDEX( r + 1 );
+         s = ( short int ) RBUFINDEX( s + 1 );
+         r = ( short int ) RBUFINDEX( r + 1 );
          if( --len )
             hb_LZSSxNodeInsert( pCompr, r );
       }
